+++ packages/eas-cli/src/commands/deploy/alias/delete.ts
@@ -44,8 +44,6 @@ export default class WorkerAliasDelete extends EasCommand {
       enableJsonOutput();
     }
 
-    Log.warn('EAS Hosting is still in preview and subject to changes.');
-
     const {
       getDynamicPrivateProjectConfigAsync,
       loggedIn: { graphqlClient },