+++ packages/eas-cli/src/commands/deploy/alias/index.ts
@@ -73,8 +73,6 @@ export default class WorkerAlias extends EasCommand {
       enableJsonOutput();
     }
 
-    Log.warn('EAS Hosting is still in preview and subject to changes.');
-
     const {
       getDynamicPrivateProjectConfigAsync,
       loggedIn: { graphqlClient },