+++ packages/eas-cli/src/commands/deploy/delete.ts
@@ -45,8 +45,6 @@ export default class WorkerDelete extends EasCommand {
       enableJsonOutput();
     }
 
-    Log.warn('EAS Hosting is still in preview and subject to changes.');
-
     const {
       getDynamicPrivateProjectConfigAsync,
       loggedIn: { graphqlClient },